requirements.txt
streamlit
pandas
numpy
pickle-mixin
python-dotenv
google-generativeai
matplotlib
seaborn
